Move prompting logic into yeoman generator.
