+++ src/VSCodeExtension/templates/library/.gitignore
@@ -1,0 +1,2 @@
+# Generated or copied files
+*.csproj.txt
